refactor: Consolidate WorkItemState type and ensureActiveState utility
- Move WorkItemState interface to shared types file
- Extract ensureActiveState utility to types/index.ts
- Fix useMemo → useEffect for ticket sync in KanbanBoard
- Remove duplicate code from KanbanBoard and TicketDetail components

Addresses Copilot review comments on PR #169.
Test coverage deferred to #180.
